server/server/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.5
decorator==5.1.1
Django==4.2.7
django-redis==5.4.0
djangorestframework==3.14.0
drf-nested-routers==0.93.4
drf-yasg==1.21.7
et-xmlfile==1.1.0
executing==2.0.1
idna==3.4
inflection==0.5.1
ipython==8.17.2
jedi==0.19.1
Markdown==3.5.1
matplotlib-inline==0.1.6
openpyxl==3.1.2
packaging==23.2
parso==0.8.3
Pillow==10.1.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.13.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
wcwidth==0.2.9
